refactor: export a map of opcodes instead of an array
