+++ src/chains/arbitrum/vm/opcodes/index.ts
@@ -1,10 +1,8 @@
-import { opcodes as mainnetArithmeticOpcodes } from '@/chains/mainnet/vm/opcodes/arithmetic';
-import { opcodes as mainnetMemoryOpcodes } from '@/chains/mainnet/vm/opcodes/memory';
-import { Opcode } from '@/chains/types';
-import { opcodes as blockOpcodes } from './block';
+import { opcodes as mainnetOpcodes } from '@/chains/mainnet/vm/opcodes';
+import { OpcodeMap } from '@/chains/types';
+import { number } from './block/number';
 
-const modifiedOpcodes: Opcode[] = [...blockOpcodes];
-
-const mainnetOpcodes: Opcode[] = [...mainnetArithmeticOpcodes, ...mainnetMemoryOpcodes];
-
-export const opcodes: Opcode[] = [...modifiedOpcodes, ...mainnetOpcodes];
+export const opcodes: OpcodeMap = {
+  ...mainnetOpcodes,
+  0x43: number,
+};
