--- src/chains/mainnet/vm/opcodes/arithmetic/index.ts
@@ -1,4 +1,0 @@
-import { Opcode } from '@/chains/types';
-import { add } from './add';
-
-export const opcodes: Opcode[] = [add];
