--- src/chains/mainnet/vm/opcodes/block/index.ts
@@ -1,5 +1,0 @@
-import { Opcode } from '@/chains/types';
-import { blockhash } from './blockhash';
-import { number } from './number';
-
-export const opcodes: Opcode[] = [blockhash, number];
