+++ src/chains/mainnet/vm/opcodes/index.ts
@@ -1,6 +1,12 @@
-import { Opcode } from '@/chains/types';
-import { opcodes as arithmeticOpcodes } from './arithmetic';
-import { opcodes as blockOpcodes } from './block';
-import { opcodes as memoryOpcodes } from './memory';
+import { OpcodeMap } from '@/chains/types';
+import { add } from './arithmetic/add';
+import { blockhash } from './block/blockhash';
+import { number } from './block/number';
+import { mstore } from './memory/mstore';
 
-export const opcodes: Opcode[] = [...arithmeticOpcodes, ...blockOpcodes, ...memoryOpcodes];
+export const opcodes: OpcodeMap = {
+  0x1: add,
+  0x40: blockhash,
+  0x43: number,
+  0x52: mstore,
+};
