--- src/chains/mainnet/vm/opcodes/memory/index.ts
@@ -1,4 +1,0 @@
-import { Opcode } from '@/chains/types';
-import { mstore } from './mstore';
-
-export const opcodes: Opcode[] = [mstore];
