--- src/chains/optimism/vm/opcodes/block/index.ts
@@ -1,9 +1,0 @@
-import { blockhash } from '@/chains/mainnet/vm/opcodes/block/blockhash';
-import { Opcode } from '@/chains/types';
-import { number } from './number';
-
-const modifiedOpcodes: Opcode[] = [number];
-
-const mainnetOpcodes: Opcode[] = [blockhash];
-
-export const opcodes: Opcode[] = [...modifiedOpcodes, ...mainnetOpcodes];
